chore: moved test to single file
